backend/requirements.txt
fastapi[standard]>=0.115.0,<1.0.0
SQLAlchemy>=2.0.0,<3.0.0
asyncpg>=0.29.0,<1.0.0
redis[hiredis]>=5.0.0,<6.0.0
aiofiles>=24.0.0,<25.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
uvicorn[standard]>=0.30.0,<1.0.0
alembic>=1.13.0,<2.0.0
bcrypt>=4.0.0,<5.0.0
pydantic-settings>=2.0.0,<3.0.0
pydantic[email]>=2.0.0,<3.0.0

# ── Dev / Test ────────────────────────────────────────────────────────
pytest>=8.0.0,<9.0.0
pytest-asyncio>=0.23.0,<1.0.0
